修改一些处理逻辑：
* filterStmt 生成时直接进行 Expression bind

* value compare 增加 cast_to 逻辑

+++ src/observer/sql/operator/order_by_physical_operator.cpp
@@ -42,7 +42,7 @@ RC OrderByPhysicalOperator::next() {
     }
   }
   idx_++;
-  if(idx_ == tuples_.size()){
+  if(idx_ == static_cast<int64_t>(tuples_.size())){
     return RC::RECORD_EOF;
   }
   return RC::SUCCESS;
